pyproject.toml
[project]
name = "geovision"
version = "0.1.0"
description = "Conversational geospatial image analysis with multi-agent YOLO inference on Modal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-langgraph>=0.0.28",
    "copilotkit>=0.1.83",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.12",
    "langgraph>=1.0.10",
    "modal>=0.73.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
